test(package): isolate pack fixtures from repository build
